% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vulnerability_metric.R
\name{agg_by_uf}
\alias{agg_by_uf}
\title{This function groups states based on average temperature (vulnerability)}
\usage{
agg_by_uf(year = 2016)
}
\arguments{
\item{year}{is a interger value}
}
\value{
The four groups of vulnerabilities
}
\description{
This function groups states based on average temperature (vulnerability)
}
\examples{
agg_by_uf(year = 2016)

}
