% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aneel_decfec_transform_wider.R
\name{aneel_decfec_transform_wider}
\alias{aneel_decfec_transform_wider}
\title{Aneel - DEC e FEC trasnform wider}
\usage{
aneel_decfec_transform_wider(data_dir = "data_raw/decfec")
}
\arguments{
\item{dir}{Directory where the data from Aneel - Indicadores de Continuidade
DEC e FEC data has been downloaded.}
}
\value{
A dataset transformed.
}
\description{
This function read the data from Aneel - Indicadores de Continuidade DEC e FEC
and transforms in a wider dataframe.
}
\examples{
\dontrun{
  aneel_decfec_wider <- aneel_decfec_transform_wider(
    dir = "data_aneel"
  )
}
}
