% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aneel_perdas_transform_wider.R
\name{aneel_perdas_transform_wider}
\alias{aneel_perdas_transform_wider}
\title{ANEEL - PERDAS transform wider}
\usage{
aneel_perdas_transform_wider(dir = "data_raw/aneel_perdas/")
}
\arguments{
\item{dir}{Directory where the data from Aneel - Relatório de Perdas de Energia compared
data has been downloaded.}
}
\value{
A dataframe
}
\description{
This function read the data from Aneel - Relatório de Perdas de Energia compared
data and transform in wider version.
}
\examples{
\dontrun{
  aneel_perdas_wider <- aneel_perdas_transform_wider(
    dir = "data_aneel"
  )
}
}
