% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aneel_samp_transform_wider.R
\name{aneel_samp_transform_wider}
\alias{aneel_samp_transform_wider}
\title{ANEEL SAMP Transform Wider}
\usage{
aneel_samp_transform_wider(dir = "data_raw/aneel_samp", ano = NULL)
}
\arguments{
\item{dir}{Directory where the raw data of ANEEL - SAMP was downloaded.}

\item{ano}{Year of the raw data. If `NULL` will transform the raw data for all
years. As default uses `NULL`.}
}
\value{
A dataset transformed.
}
\description{
This function reads the raw data from ANEEL - SAMP and transforms in a wider
dataframe. Each line represents a subclass consumer of a energy distributor by
UF and year.
}
\examples{
\dontrun{
  aneel_samp_wider <- aneel_samp_transform_wider(
    dir = "data_raw/aneel_samp",
    ano = 2022
  )
}
}
