% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preco_glp_winder.R
\name{anp_precoglp_transform_wider}
\alias{anp_precoglp_transform_wider}
\title{ANP - Preço de GLP trasnform wider}
\usage{
anp_precoglp_transform_wider(
  dir = "data_raw/preco_glp/",
  file_name = "preco_glp_2004_2024.rds"
)
}
\arguments{
\item{dir}{Directory where the data from ANP - Preço de GLP data has been downloaded. ("data_raw/preco_glp/")}

\item{file_name}{.rds File name where the data from ANP - Preço de GLP data has been saved. (default: "preco_glp_2004_2024.rds")}
}
\value{
A dataframe
}
\description{
This function read the data from ANP - Preço de GLP
and transforms in a wider dataframe.
}
\examples{
\dontrun{
  anp_precoglp_transform_wider <- anp_precoglp_transform_wider()
}
}
