% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixa_aneel_perdas.R
\name{baixa_aneel_perdas}
\alias{baixa_aneel_perdas}
\title{Exporta dados de perdas energéticas da ANEEL via Selenium}
\usage{
baixa_aneel_perdas()
}
\description{
Automatiza o processo de exportação dos dados de perdas energéticas do portal da ANEEL.
}
