% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixa_aneel_samp.R
\name{baixa_aneel_samp}
\alias{baixa_aneel_samp}
\title{Baixa ANEEL - SAMP (Balanço Energético)}
\usage{
baixa_aneel_samp(destination_dir = "data_raw/aneel_samp")
}
\arguments{
\item{destination_dir}{O caminho do diretório base onde os dados serão salvos. Por padrão,
é 'data_raw/aneel_samp/'.}
}
\value{
A função não retorna um valor explicitamente, mas como efeito colateral,
arquivos são baixados e extraídos nos diretórios especificados.
}
\description{
Essa função é responsável por baixar os dados de Balanço Energético (SAMP) da ANEEL.
Os arquivos são baixados diretamente do repositório dos Dados Abertos da ANEEL
e salvo no diretório destinado.
}
\examples{
\dontrun{
  baixa_aneel_samp(
    destination_dir = 'data_raw/aneel_samp'
  )
}
}
