% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixa_cadunico_amostral.R
\name{baixa_cadunico_amostral}
\alias{baixa_cadunico_amostral}
\title{Baixa CadÚnico}
\usage{
baixa_cadunico_amostral(destination_dir = "data_raw/cadunico/", ano = NULL)
}
\arguments{
\item{destination_dir}{O caminho do diretório base onde os dados serão salvos. Por padrão,
é 'data_raw/cadunico/'.}

\item{ano}{Ano de referência dos dados a serem baixados.}
}
\value{
A função não retorna um valor explicitamente, mas como efeito colateral,
arquivos são baixados e extraídos nos diretórios especificados.
}
\description{
Essa função é responsável por baixar os dados do amostrais do Cadastro Único do
Ministério do Desenvolvimento e Assistência Social, Família e Combate a Fome.
Os arquivos são baixados diretamente do repositório do GOV e salvo no diretório destinado.
}
\examples{
\dontrun{
  baixa_cadunico_amostral(
    destination_dir = 'data_raw/cadunico'
  )
}
}
