% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixa_censo.R
\name{baixa_censo}
\alias{baixa_censo}
\title{Baixa e Prepara Dados Agregados Preliminares do Censo IBGE 2022}
\usage{
baixa_censo(destination_dir = "data-raw/censo-ibge-2022/")
}
\arguments{
\item{destination_dir}{Diretório de destino onde os arquivos baixados serão salvos e extraídos.
Por padrão, é 'data-raw/censo-ibge-2022/'.}
}
\value{
Um tibble com os dados lidos do arquivo CSV dos agregados preliminares por setores censitários.
Os nomes das colunas são limpos para facilitar a manipulação dos dados.
}
\description{
Esta função automatiza o download dos dados agregados por setores censitários preliminares
do Censo Demográfico 2022 do IBGE. Ela baixa o arquivo zip do website do IBGE, extrai o conteúdo
no diretório especificado e carrega o arquivo CSV resultante, limpando os nomes das colunas.
}
