% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixa_epe_simples.R
\name{baixa_epe_simples}
\alias{baixa_epe_simples}
\title{Baixa EPE Simples}
\usage{
baixa_epe_simples(destination_dir = "data_raw/epe_simples")
}
\arguments{
\item{destination_dir}{O caminho do diretório base onde os dados serão salvos. Por padrão,
é 'data_raw/epe_simples/'.}
}
\value{
A função não retorna um valor explicitamente, mas como efeito colateral,
arquivos são baixados e extraídos nos diretórios especificados.
}
\description{
Essa função é responsável por baixar os dados do Anuário Estatístico de Energia Elétrica
da EPE. Os arquivos são baixados diretamente do repositório dos Dados Abertos da EPE
e salvo no diretório destinado.
}
\examples{
\dontrun{
  baixa_epe_simples(
    destination_dir = 'data_raw/epe_simples'
  )
}
}
