% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixa_mds_auxgas.R
\name{baixa_mds_auxgas}
\alias{baixa_mds_auxgas}
\title{Baixa MDS - Auxílio Gás}
\usage{
baixa_mds_auxgas(destination_dir = "data_raw/mds_auxgas")
}
\arguments{
\item{destination_dir}{O caminho do diretório base onde os dados serão salvos. Por padrão,
é 'data_raw/mds_auxgas/'.}
}
\value{
A função não retorna um valor explicitamente, mas como efeito colateral,
arquivos são baixados e extraídos nos diretórios especificados.
}
\description{
Essa função é responsável por baixar os dados do Programa Auxílio Gás para Brasileiros
do Ministério do Desenvolvimento e Assistência Social, Família e Combate a Fome (MDS)
entre os anos de 2021 a 2024. Os arquivos são baixados diretamente do repositório dos
Dados Abertos do GOV e salvo no diretório destinado.
}
\examples{
\dontrun{
  baixa_mds_auxgas(
    destination_dir = 'data_raw/mds_auxgas'
  )
}
}
