% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixa_pibm.R
\name{baixa_pibm}
\alias{baixa_pibm}
\title{Baixa IBGE - PIB Municipal}
\usage{
baixa_pibm(destination_dir = "data_raw/pibm/")
}
\arguments{
\item{destination_dir}{O caminho do diretório base onde os dados serão salvos. Por padrão,
é 'data_raw/pibm/'. A função criará subdiretórios para os anos de 2002-2009 e 2010-2021 dentro deste
diretório e organizará os arquivos baixados e extraídos nessas pastas.}
}
\value{
A função não retorna um valor explicitamente, mas como efeito colateral,
arquivos são baixados e extraídos nos diretórios especificados.
}
\description{
Essa função é responsável por baixar os dados de Produto Interno Bruto dos Municípios
do IBGE para os anos entre 2002 a 2021. Os arquivos são baixados diretamente
dos repositórios do IBGE e extraídos em diretórios específicos dentro do diretório
de destino fornecido.
}
\examples{
\dontrun{
  baixa_pibm(
   destination_dir = 'data_raw/pibm/'
  )
}
}
