% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixa_pof.R
\name{baixa_pof}
\alias{baixa_pof}
\title{Esta função é responsável por baixar e extrair os arquivos de microdados e documentação
da Pesquisa de Orçamentos Familiares (POF) para os anos de 2018 e 2009. Os arquivos são
baixados diretamente dos repositórios do IBGE e extraídos em diretórios específicos
dentro do diretório de destino fornecido.}
\usage{
baixa_pof(destination_dir = "data-raw/POF/")
}
\arguments{
\item{destination_dir}{O caminho do diretório base onde os dados serão salvos. Por padrão,
é 'data-raw/POF/'. A função criará subdiretórios para os anos de 2018 e 2009 dentro deste
diretório e organizará os arquivos baixados e extraídos nessas pastas.}
}
\value{
A função não retorna um valor explicitamente, mas como efeito colateral,
arquivos são baixados e extraídos nos diretórios especificados.
}
\description{
Esta função é responsável por baixar e extrair os arquivos de microdados e documentação
da Pesquisa de Orçamentos Familiares (POF) para os anos de 2018 e 2009. Os arquivos são
baixados diretamente dos repositórios do IBGE e extraídos em diretórios específicos
dentro do diretório de destino fornecido.
}
\examples{
baixa_pof() # Baixa e extrai os dados da POF para o diretório padrão 'data-raw/POF/'
baixa_pof('meu_diretorio_de_dados/') # Especifica um diretório diferente para o download e extração dos dados

}
