% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixa_preco_glp.R
\name{baixa_preco_glp}
\alias{baixa_preco_glp}
\title{Baixa Preço do GLP}
\usage{
baixa_preco_glp(
  destination_dir = "data_raw/preco_glp",
  ano_inicial = 2004,
  ano_final = 2024
)
}
\arguments{
\item{destination_dir}{O caminho do diretório base onde os dados serão salvos. Por padrão,
é 'data_raw/preco_glp'.}

\item{ano_inicial}{Ano de interesse dos dados a serem baixados}

\item{ano_final}{Ano de interesse dos dados a serem baixados}
}
\value{
A função não retorna um valor explicitamente, mas como efeito colateral,
arquivos são baixados e extraídos nos diretórios especificados.
}
\description{
Essa função é responsável por baixar os dados do Preço do GLP segundo a Agência Nacional
do Petróleo, Gás Natural e Biocombustivéis (ANP). Os arquivos são baixados diretamente do
repositório dos Dados Abertos do GOV e salvo no diretório destinado.
}
\examples{
\dontrun{
  baixa_preco_glp(
    destination_dir = 'data_raw/preco_glp'
  )
}
}
