% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixa_visita_1_pnadc.R
\name{baixa_visita_1_pnadc}
\alias{baixa_visita_1_pnadc}
\title{Baixa Dicionário, Input e Microdados da 1ª Visita da PNADc}
\usage{
baixa_visita_1_pnadc(ano = 2018, destination_dir = "data-raw/pnadc/2018/")
}
\arguments{
\item{ano}{Ano de referência para os dados da PNADc que serão baixados.}

\item{destination_dir}{Diretório de destino onde os arquivos serão salvos.}
}
\value{
Os arquivos são baixados e salvos no diretório especificado, sem retorno explícito na função.
}
\description{
Esta função automatiza o processo de download do dicionário, arquivo de input e microdados
para a 1ª visita da Pesquisa Nacional por Amostra de Domicílios Contínua (PNADc) de um ano específico,
com suporte a variações no nome dos arquivos com base no ano. Os arquivos são salvos em um diretório especificado.
}
