% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cadunico_preprocessing.R
\name{cadunico_preprocessing}
\alias{cadunico_preprocessing}
\title{Tranform CadÚnico}
\usage{
cadunico_preprocessing(
  cadunico_familias_raw_data,
  cadunico_pessoas_raw_data,
  ano = "2022"
)
}
\arguments{
\item{cadunico_familias_raw_data}{A table that contains data of CadÚncio Famílias.}

\item{cadunico_pessoas_raw_data}{A table that contains data of CadÚncio Pessoas.}

\item{ano}{Year of the raw data.}
}
\value{
A DataFrame of selected and transformed columns of CadÚnico database by
year. Each row represents a household.
}
\description{
Apply data treatment and transformation in dataset of CadÚnico Famílias and
CadÚnico Pessoas and return a DataFrame with columns of interest by household.
}
\examples{
\dontrun{
  cadunico_transformed <- cadunico_preprocessing(
    cadunico_familias_raw_data = base_familias2022_tratada,
    cadunico_pessoas_raw_data = base_pessoas2022_tratada,
    ano = "2022"
  )
}
}
