% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_indicators.R
\name{calculate_indicators}
\alias{calculate_indicators}
\title{Calculate Indicators with Conditioning and Indicator Columns}
\usage{
calculate_indicators(
  dataset,
  determinant_columns,
  indicator_columns,
  with_filters,
  database,
  year = "2018",
  time_period = "year"
)
}
\arguments{
\item{dataset}{dataset indicadores}

\item{determinant_columns}{A character vector of column names to be pivoted as conditioning factors}

\item{indicator_columns}{A character vector of column names to be pivoted as indicators.}
}
\description{
Calculate Indicators with Conditioning and Indicator Columns
}
\examples{

determinantes_nome <- df_indicadores \%>\%
select(DC01:DD13) \%>\%
names()

indicadores_nome <- df_indicadores \%>\%
select(IA0101:IA0105) \%>\%
names()

calculate_indicators(df_indicadores,
determinant_columns = determinantes_nome,
indicator_columns = indicadores_nome)

to-do: fazer teste para garantir que os resultados finais
#o valor que deve aparecer no ref_determinante_niveis_n_ponderado:69017704
dataset  |> summarise(.by = c(UF, DC01), s = sum(PESO_FINAL)) |> summarise(sum(s))
além de garantir que os percentuais finais batam com os valores pré pivotagem
dataset  |> summarise(.by = c(UF, DC01, IA0101 ), s = sum(PESO_FINAL)) |>  arrange(UF, DC01, IA0101)

}
