% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carregar_para_sql.R
\name{carregar_para_sql}
\alias{carregar_para_sql}
\title{Envia Dados para o SQL Server}
\usage{
carregar_para_sql(dataframe, nome_tabela, servidor, database, porta)
}
\arguments{
\item{dataframe}{Caminho para o arquivo CSV, separado por ";" que contém os dados a serem enviados para o banco de dados.}

\item{nome_tabela}{Nome da tabela dentro do banco de dados onde os dados serão inseridos.}

\item{servidor}{Endereço do servidor SQL Server.}

\item{database}{Nome do banco de dados no servidor.}

\item{porta}{porta de conexão com o servidor.}
}
\description{
Esta função verifica o sistema operacional, configura a localidade apropriada e utiliza duas funções auxiliares para conectar ao SQL Server e carregar dados de um dataframe especificado para uma tabela do banco de dados.
}
\details{
A função primeiro verifica o sistema operacional para ajustar a localidade para leitura de arquivos com caracteres especiais.
Em seguida, usa a função `conectar_sql_server` para estabelecer uma conexão com o banco de dados e a função `carregar_para_sql` para enviar os dados do dataframe para a tabela especificada no banco de dados.
A função `carregar_para_sql` lê o arquivo CSV, usando a função readr::read_csv2, insere os dados na tabela do banco de dados e depois fecha a conexão.
}
