% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mme_lpt.R
\name{constroi_mme_lpt}
\alias{constroi_mme_lpt}
\title{Estrutura os dados do Luz Para Todos - Ministério de Minas e Energia}
\usage{
constroi_mme_lpt(destination_dir = "data_raw/mme_lpt")
}
\arguments{
\item{destination_dir}{Diretório onde os dados do MME - LPT serão lidos. Caso
os dados não estejam presentes no diretório indicado, os dados serão baixados.}
}
\value{
Uma tabela contedos os dados de vulnerabilidade da base do MME - LPT
}
\description{
Essa função é responsável por estruturar os dados da base Luz Para Todos (LPT) do
Ministério de Minas e Energia em uma tabela contendo os dados sumarizados que
contém as métricas de vulnerabilidade na base LPT
}
\examples{
\dontrun{
  constroi_mme_lpt(destination_dir = 'data_raw/mme_lpt')
}
}
