% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constroi_pnad.R
\name{constroi_pnad}
\alias{constroi_pnad}
\title{Constrói um Dataset Consolidado da PNADc}
\usage{
constroi_pnad(diretorio, ano)
}
\arguments{
\item{diretorio}{O caminho do diretório onde os arquivos de microdados da PNADc estão
localizados. Este diretório deve conter tanto os arquivos de microdados quanto os arquivos
de input e dicionários necessários para a leitura e rotulagem dos dados.}

\item{ano}{O ano dos dados da PNADc que serão processados. Embora a função esteja preparada
para tratar especificidades de diferentes anos, atualmente há uma diferenciação específica
no processamento entre o ano de 2022 e outros anos.}
}
\value{
Um tibble que representa a tabela consolidada combinando dados das visitas 1 e 5
da PNADc, incluindo variáveis selecionadas e processadas conforme especificado. A forma
como os dados das duas visitas são combinados depende do ano especificado.
}
\description{
Esta função combina os microdados da primeira e quinta visita da Pesquisa Nacional por
Amostra de Domicílios Contínua (PNADc) para um ano específico, aplicando rotulagem
e selecionando variáveis de interesse. A função facilita o processo de preparação
dos dados para análises subsequentes.
}
