% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constroi_pnad_tabelas.R
\name{constroi_pnad_tabelas}
\alias{constroi_pnad_tabelas}
\title{Constrói Datasets Separados da PNADc}
\usage{
constroi_pnad_tabelas(diretorio, ano)
}
\arguments{
\item{diretorio}{O caminho do diretório onde os arquivos de microdados da PNADc estão
localizados. Este diretório deve conter tanto os arquivos de microdados quanto os arquivos
de input e dicionários necessários para a leitura e rotulagem dos dados.}

\item{ano}{O ano dos dados da PNADc que serão processados. Embora a função esteja preparada
para tratar especificidades de diferentes anos, atualmente há uma diferenciação específica
no processamento entre o ano de 2022 e outros anos.}
}
\value{
Uma lista com dois tibbles, onde cada tibble representa os dados da visita 1 e 5
da PNADc, incluindo variáveis selecionadas e processadas conforme especificado.
}
\description{
Esta função lê e processa os microdados da primeira e quinta visita da Pesquisa Nacional por
Amostra de Domicílios Contínua (PNADc) para um ano específico, aplicando rotulagem e selecionando
variáveis de interesse. Os dados de cada visita são retornados em uma lista separada.
}
