% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constroi_pof.R
\name{constroi_pof}
\alias{constroi_pof}
\title{Constrói um Dataset Consolidado da POF}
\usage{
constroi_pof(diretorio = "data-raw/POF/2018/")
}
\arguments{
\item{diretorio}{O caminho do diretório onde os arquivos de dados da POF estão localizados. Por padrão,
é 'data-raw/POF/2018/'. Espera-se que este diretório contenha subdiretórios e arquivos específicos conforme
a estrutura utilizada pela POF para armazenamento de dados e metadados.}
}
\value{
Um tibble que representa a tabela única consolidada com os dados da POF, incluindo variáveis selecionadas
e processadas conforme especificado.
}
\description{
Esta função automatiza o processo de leitura e combinação de múltiplos datasets da Pesquisa de Orçamentos Familiares (POF)
para o ano especificado, baseando-se em um conjunto de variáveis de interesse. Ela realiza a leitura de arquivos
texto baseada em um dicionário de variáveis, seleciona variáveis específicas, limpa nomes de colunas, e junta múltiplos
datasets em uma única tabela.
}
