% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constroi_pof_tabelas.R
\name{constroi_pof_tabelas}
\alias{constroi_pof_tabelas}
\title{Constrói Datasets da POF}
\usage{
constroi_pof_tabelas(diretorio = "data-raw/POF/2018/")
}
\arguments{
\item{diretorio}{O caminho do diretório onde os arquivos de dados da POF estão localizados. Por padrão,
é 'data-raw/POF/2018/'. Espera-se que este diretório contenha subdiretórios e arquivos específicos conforme
a estrutura utilizada pela POF para armazenamento de dados e metadados.}
}
\value{
Uma lista de tibbles, onde cada tibble representa um dataset da POF, incluindo variáveis selecionadas
e processadas conforme especificado.
}
\description{
Esta função automatiza o processo de leitura de múltiplos datasets da Pesquisa de Orçamentos Familiares (POF)
para o ano especificado, baseando-se em um conjunto de variáveis de interesse. Ela realiza a leitura de arquivos
texto baseada em um dicionário de variáveis, seleciona variáveis específicas, limpa nomes de colunas, e retorna
uma lista de tibbles, onde cada tibble representa um dataset da POF.
}
