% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_aneel_decfec_vulnerability.R
\name{create_aneel_decfec_vulnerability}
\alias{create_aneel_decfec_vulnerability}
\title{Creates Aneel - DEC e FEC Vulnerability}
\usage{
create_aneel_decfec_vulnerability(aneel_decfec_wider)
}
\arguments{
\item{aneel_decfec_wider}{Aneel - DEC e FEC data transformed in wider data.}
}
\value{
A dataframe
}
\description{
This function transform the aneel_decfec_wider data and summarizes all vulnerabilities
at UF, region, and country view per stats.
}
\examples{
\dontrun{
  aneel_decfec_summarized_by_vulnerability <- create_aneel_perdas_vulnerability(
    dir = "data_aneel"
  )
}
}
