% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_aneel_samp_vulnerability.R
\name{create_aneel_samp_vulnerability}
\alias{create_aneel_samp_vulnerability}
\title{Creates ANEEL - SAMP Vulnerability}
\usage{
create_aneel_samp_vulnerability(aneel_samp_wider)
}
\arguments{
\item{aneel_samp_wider}{A dataframe of ANEEL - SAMP transformed in a wider version}
}
\value{
A dataframe containing the summarization of the vulnerabilities in
geographic per year for each vulnerability
}
\description{
This function reads the wider data version of ANEEL - SAMP and summarizes all
vulnerabilities in a uf, region, country per year for each vulnerability.
}
\examples{
\dontrun{
  aneel_samp_summarized_by_vulnerabilities <- create_aneel_samp_vulnerability(
    aneel_samp_wider = aneel_samp_wider
  )
}
}
