% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_anp_precoglp_vulnerability.R
\name{create_anp_precoglp_vulnerability}
\alias{create_anp_precoglp_vulnerability}
\title{Creates ANP - PRECOGLP Vulnerability}
\usage{
create_anp_precoglp_vulnerability(
  dir = "data_raw/preco_glp/",
  file_name = "preco_glp_2004_2024.rds"
)
}
\arguments{
\item{dir}{Directory where the data from ANP - Preço de GLP data has been downloaded. ("data_raw/preco_glp/")}

\item{file_name}{.rds File name where the data from ANP - Preço de GLP data has been saved. (default: "preco_glp_2004_2024.rds")}
}
\value{
A dataframe
}
\description{
This function read the data from ANP - Preço do GLP compared
data and summarizes all vulnerabilities at UF, region, and country view per stats.
}
\examples{
\dontrun{
  anp_precoglp_summarized_by_vulnerability <- create_anp_precoglp_vulnerability()
}
}
