% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_indicator_data.R
\name{create_indicator_data}
\alias{create_indicator_data}
\title{Create Indicator Data}
\usage{
create_indicator_data(
  path,
  database,
  year,
  indicator,
  export_path = "ETL/output/longer_indicators/"
)
}
\arguments{
\item{path}{A character string specifying the file path to the dataset in `.rds` format.}

\item{database}{A character string indicating the database used for the data processing.}

\item{year}{An integer representing the year of the data.}

\item{indicator}{A character vector specifying the indicator(s) to be calculated.}

\item{export_path}{A character string specifying the path to export the resulting dataset.}
}
\value{
A data frame containing the aggregated metrics, with columns for
various levels and determinants, such as per household or individual,
geographic levels (UF, REG, BRA), and with or without determinants.
}
\description{
Generates a dataset with aggregated metrics for a given indicator across
multiple levels and with or without determinants.
}
\details{
This function performs multiple aggregations based on the indicator specified.
It reads the dataset from the specified path, then applies 12 types of aggregations,
which include combinations of:
- Aggregation level: per household vs. individual
- Geographic level: UF, REG, BRA
- Determinants: with or without determinant filters
}
\examples{
\dontrun{
create_indicator_data(
  path = "path/to/datafile.rds",
  database = "SampleDatabase",
  year = 2022,
  indicator = c("indicator1", "indicator2")
)
}

}
