% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_inmet_vulnerability.R
\name{create_inmet_vulnerability}
\alias{create_inmet_vulnerability}
\title{Create INMET Vulnerabilities}
\usage{
create_inmet_vulnerability()
}
\value{
A dataframe.
}
\description{
This function reads the data from INMET (Instituto Nacional de Meteriologia)
and summarizes all vulnerabilies at UF, region and country view per stats.
}
\examples{
\dontrun{
  inmet_summarized_by_vulnerability <- create_inmet_vulnerability()
}
}
