% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ipea_vulnerability.R
\name{create_ipea_vulnerability}
\alias{create_ipea_vulnerability}
\title{Create IPEA Vulnerability}
\usage{
create_ipea_vulnerability()
}
\value{
A dataframe.
}
\description{
This function read the data from IPEA API for IDHM, IDHM Longevidade, IDHM Renda,
IDHM Educação and summarizes at UF, region and country view per stats.
}
\examples{
\dontrun{
  ipea_summarized_by_vulnerability <- create_ipea_vulnerability()
}
}
