% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_longer_summarized.R
\name{create_longer_summarized}
\alias{create_longer_summarized}
\title{Create Longer Summarized by Determinants}
\usage{
create_longer_summarized(
  dir = "./ETL_pipeline/data/data-output",
  exdir = "./ETL_pipeline/data/data-output"
)
}
\arguments{
\item{dir}{Directory where the longer dataset is located. As default uses
'./ETL_pipeline/data/data-output'.}

\item{exdir}{Export directory where the longer summarized by determinants will
be saved. As default uses './ETL_pipeline/data/data-output'.}
}
\value{
A version of the longer dataset summarized by determinants in the
export directory (exdir) selected.
}
\description{
Creates a version of the Longer dataset Summarized by Determinants
}
\examples{
\dontrun{
create_longer_summarized(
 dir = "./dataset_longer",
 exdir = "./output_dataset_longer"
)
}
First, we need to create a column that reports the volume represented by the
determinants, since in the original dataset, we only have this volume for
the levels of these determinants. Note that the information
ref_total_filtro_and_indicator_n_ponderado will be exactly the same for all
possible indicators, while it needs to be calculated within an indicator
to result in the correct quantities.
treted merged
}
