% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_vulnerabilities_summarized.R
\name{create_vulnerabilities_summarized}
\alias{create_vulnerabilities_summarized}
\title{Create Vulnerabilities Summarized}
\usage{
create_vulnerabilities_summarized(exdir = "./ETL_pipeline/data/data-output/")
}
\arguments{
\item{exdir}{Directory destination where the table will be saved.}
}
\value{
A table saved in directory destination containing vulnerabilities data
}
\description{
This function creates a summarized table containing the vulnerabilities
present in INMET and IPEA data
}
\examples{
\dontrun{
 create_vulnerabilities_summarized(
 exdir = "vulnerabilidades/"
 )
}
}
