% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creates_validation_report.R
\name{creates_validation_report}
\alias{creates_validation_report}
\title{Create Validation Report}
\usage{
creates_validation_report(export_path = "ETL_pipeline/data/data-output/")
}
\arguments{
\item{export_path}{A character string specifying the directory in
which the longer and wider datasets are stored.
Default is `"ETL_pipeline/data/data-output/"`.}
}
\value{
None. The function saves the report as a HTML file
in the specified `export_path`.
}
\description{
Creates a HTML report to confirm the integrity
of the wider and longer datasets.
}
\details{
This function loads each specified `.rds` file containing the
wider and longer datasets. It then invokes a .rmd script to
create a HTML report that is saved on the `export_path`.
}
\examples{
\dontrun{
creates_validation_report(
  export_path = "ETL_pipeline/data/data-output/"
)
}

}
