% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creates_wider_summarized.R
\name{creates_wider_summarized}
\alias{creates_wider_summarized}
\title{Creates Wider Summarized by Additional Stats}
\usage{
creates_wider_summarized(
  dir = "./ETL_pipeline/data/data-output/microdados-wider-rds",
  exdir = "./ETL_pipeline/data/data-output"
)
}
\arguments{
\item{dir}{Directory that contais all wider data. As default uses
'./ETL_pipeline/data/data-output/microdados-wider-rds' as directory}

\item{exdir}{Directory that will save the wider summarized. As default uses
'./ETL_pipeline/data/data-output'}
}
\value{
A RDS file of a dataset containing all Additional Stats summarized
by country, region and UF named by '_df_metrics_wider_summarized_by_EA'
}
\description{
This function read all wider data in a directory and return a dataset that
summarizes all Additional Stats by country, region and UF as RDS file.
}
\examples{
\dontrun{
creates_wider_summarized(
  dir = './wider_data',
  exdir = './output'
  )
}
}
