% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extração_aneel_decfec_dados_nacionais.R
\name{export_national_data_2004}
\alias{export_national_data_2004}
\title{Exporta dados de DEC e FEC nacionais da ANEEL via Selenium para o período de 2004 a 2013}
\usage{
export_national_data_2004(remDr)
}
\arguments{
\item{remDr}{Objeto do tipo `remoteDriver`, utilizado para controlar a página web.}
}
\value{
Nenhum valor explícito é retornado. 
A função exibe mensagens de status durante o processo 
e exporta a tabela de dados.
}
\description{
Esta função utiliza um objeto `remoteDriver` do pacote `RSelenium` para interagir 
com a página web da ANEEL e exportar os dados dos indicadores de continuidade DEC e FEC 
a nível nacional, abrangendo o período de 2004 a 2013.
}
\examples{
\dontrun{
remDr <- RSelenium::remoteDriver(...)
remDr$open()
export_national_data_2004(remDr)
}
}
