% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_aneel_decfec.R
\name{export_table}
\alias{export_table}
\title{Exporta Dados de uma Tabela Específica do portal da ANEEL via Selenium}
\usage{
export_table(remDr, nivel)
}
\arguments{
\item{remDr}{Um objeto `RemoteDriver` do pacote RSelenium, utilizado para controlar
o navegador e realizar ações de automação na página da web.}

\item{nivel}{Uma string indicando o nível da tabela a ser exportada. Os valores válidos são:
- `"Nacional"`: Tabela de dados em nível nacional.
- `"Regional_2000"`: Tabela de dados em nível regional para o período de 2000 a 2009.
- `""`: Seleciona as tabelas de dados a nível regional/uf, exceto para a tabela regional dos anos de 2000 a 2009.}
}
\value{
Nenhum valor de retorno.
O processo de exportação é iniciado diretamente no navegador, com o download dos dados realizado automaticamente.
}
\description{
A função `export_table` interage com o Selenium para localizar e exportar dados de
uma tabela específica do portal da ANEEL. Ela move o mouse até a tabela e clica no
botão "Mais Opções" e, em seguida, "Exportar dados", para baixar os dados automaticamente.
}
\details{
A função localiza a tabela específica na página, move o mouse até a posição da tabela e realiza uma sequência de
interações para visualizar o botão "Mais Opções", o qual é clicado para acessar a opção de exportação dos dados.
O tempo de espera é controlado através de `Sys.sleep()` para garantir a estabilidade e a correta execução das interações.
}
\examples{
\dontrun{
remDr <- RSelenium::remoteDriver(browserName = "firefox")
remDr$open()
remDr$setTimeout(timeout = 10000)  # Definindo o tempo de espera máximo
export_table(remDr, "Nacional")
}

}
