% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extração_aneel_decfec_dados_uf.R
\name{export_uf_data_2004}
\alias{export_uf_data_2004}
\title{Exporta dados de DEC e FEC da ANEEL por UF via Selenium para o período de 2010 a 2013}
\usage{
export_uf_data_2004(remDr)
}
\arguments{
\item{remDr}{Objeto do tipo remoteDriver, utilizado para controlar a página web.}
}
\value{
Nenhum valor explícito é retornado. 
A função exibe mensagens de status durante o processo 
e exporta a tabela de dados dos indicadores DEC e FEC a nível de UF.
}
\description{
Esta função utiliza um objeto remoteDriver do pacote RSelenium para interagir 
com a página web da ANEEL e exportar os dados dos indicadores de continuidade DEC e FEC 
a nível de UF (Unidade Federativa), abrangendo o período de 2010 a 2013.
}
\examples{
\dontrun{
remDr <- RSelenium::remoteDriver(...)
remDr$open()
export_uf_data_2004(remDr)
}
}
