% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds_auxgas_transform_wider.R
\name{mds_auxgas_transform_wider}
\alias{mds_auxgas_transform_wider}
\title{MDS - Auxílio Gás transform wider}
\usage{
mds_auxgas_transform_wider(dir = "data_raw/mds_auxgas")
}
\arguments{
\item{dir}{Directory where the data from MDS - Auxílio Gás has been downloaded.}
}
\value{
A dataframe
}
\description{
This function read the data from MDS - Auxílio Gás and transforms in a wider
version.
}
\examples{
\dontrun{
  mds_auxgas_wider <- mds_auxgas_transform_wider(
    dir = "data_mds_auxgas"
  )
}
}
