% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_aneel_decfec.R
\name{navigate}
\alias{navigate}
\title{Navega até a Página de Indicadores DEC/FEC do portal da ANEEL usando Selenium}
\usage{
navigate(remDr)
}
\arguments{
\item{remDr}{Um objeto `RemoteDriver` do pacote RSelenium, utilizado para controlar
o navegador e realizar ações de automação na página da web.}
}
\value{
Nenhum valor de retorno.
A função carrega a página dos indicadores DEC/FEC no navegador controlado pelo Selenium e
entra no `iframe` correspondente, permitindo a interação com os dados exibidos.
}
\description{
A função `navigate` acessa a página dos indicadores DEC/FEC no portal da ANEEL e
entra no `iframe` necessário para permitir a interação com os dados da página.
}
\details{
A função navega até a URL dos indicadores DEC/FEC, localiza o `iframe` dentro da página e realiza a troca de contexto
para o `iframe`, permitindo a interação com os dados exibidos na página. O tempo de espera é usado para garantir que
a página carregue corretamente e o `iframe` esteja pronto para interação.
}
\examples{
\dontrun{
remDr <- RSelenium::remoteDriver(browserName = "firefox")
remDr$open()
remDr$setTimeout(timeout = 10000)  # Definindo o tempo de espera máximo
navigate(remDr)
}

}
