% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vulnerability_metric.R
\name{read_temp_file}
\alias{read_temp_file}
\title{Title}
\usage{
read_temp_file(path = "./data/cdd_hdd_uf_2014_2023_dashboard.rds")
}
\arguments{
\item{path}{Path of .rds with the temperatures}
}
\value{
dataframe with the rds file or NULL
}
\description{
Title
}
