% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_wider.R
\name{save_wider}
\alias{save_wider}
\title{Save Wider}
\usage{
save_wider(
  data_wider,
  database_name,
  database_year,
  exdir = list(rds = "ETL_pipeline/data/data-output/microdados-wider-rds/", csv =
    "ETL_pipeline/data/data-output/microdados-wider-csv/")
)
}
\arguments{
\item{data_wider}{Database with metrics calculated. Output of calculate metrics
functions}

\item{database_name}{Name of the database. "pof" or "pnad"}

\item{database_year}{Year of the database}

\item{exdir}{Export directory destination}
}
\value{
Saves wider dataset in the export directory
}
\description{
Save Wider
}
\examples{
\dontrun{
save_wider(
  data_wider = pnad2019_wider,
  database_name = "pnad",
  database_year = 2019,
  exdir = list(rds = "ETL_pipeline/data/data-output/microdados-wider-rds/",
               csv = "ETL_pipeline/data/data-output/microdados-wider-csv/")
)
}
}
