% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_database.R
\name{transform_database}
\alias{transform_database}
\title{Transform Database}
\usage{
transform_database(
  database,
  database_name,
  database_year,
  exdir = "ETL_pipeline/data/data-log"
)
}
\arguments{
\item{database}{Raw-database object. A list for pof and a dataframe for pnad.}

\item{database_name}{Name of the database. "pof" or "pnad"}

\item{database_year}{Year of the database.}

\item{exdir}{Directory where the rds file will be saved. As default uses
'ETL_pipeline/data/data-log'.}
}
\value{
A rds file with the transformed database in the export directory
destination
}
\description{
This function receives a object (list or dataframe) with the raw database,
the output of the constroi_pof_tabelas or constroi_pnad_tabelas functions,
and returns a rds file saved in the export directory destination that contains
the database with selected columns and transformation needed for each database.
}
\examples{
\dontrun{
transform_database(
  database = pof_2009,
  database_name = "pof",
  database_year = 2009,
  exdir = './data-output'
)
}
}
