% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_pnad.R
\name{transform_pnad}
\alias{transform_pnad}
\title{Tranform PNADCA}
\usage{
transform_pnad(pnad_data)
}
\arguments{
\item{pnad_data}{A table that contains data of PNADCA Visita 1.}
}
\value{
A DataFrame of selected and transformed columns of PNADC Visita 1 by
year. Each row represents a household.
}
\description{
Apply data treatment and transformation in dataset of PNADCA Visita 1 and
return a DataFrame with columns of interest by household.
}
\examples{
\dontrun{
transform_pnad(
  pnad_data = pnad2019
)
}
}
