% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_pof.R
\name{transform_pof}
\alias{transform_pof}
\title{Transform POF}
\usage{
transform_pof(pof_data, pof_year)
}
\arguments{
\item{pof_data}{A list vector of POF containing each table of POF.}

\item{pof_year}{A numeric vector that identifies the year of the POF.}
}
\value{
A DataFrame of selected columns of every table of interest in POF
by year. Each row represents a 'Unidade de Consumo'.
}
\description{
Apply data treatment and transformation in tables of interest in POF and
returning a dataset with selected tables and columns of interest by Unidades
de consumo.
}
\examples{
\dontrun{
transform_pof(
  pof_data = pof2009,
  2009
)
}

}
