% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_aneel_decfec_vulnerability.R
\name{create_aneel_decfec_vulnerability}
\alias{create_aneel_decfec_vulnerability}
\title{Creates Aneel - DEC e FEC Vulnerability}
\usage{
create_aneel_decfec_vulnerability(data_dir = "data_raw/decfec")
}
\arguments{
\item{dir}{Directory where the data from Aneel - Indicadores de Continuidade
DEC e FEC data has been downloaded.}
}
\value{
A dataframe
}
\description{
This function read the data from Aneel - Indicadores de Continuidade DEC e FEC
and summarizes all vulnerabilities at UF, region, and country view per stats.
}
\examples{
\dontrun{
  aneel_decfec_summarized_by_vulnerability <- create_aneel_perdas_vulnerability(
    dir = "data_aneel"
  )
}
}
