% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_aneel_perdas_vulnerability.R
\name{create_aneel_perdas_vulnerability}
\alias{create_aneel_perdas_vulnerability}
\title{Creates Aneel - Perdas Vulnerability}
\usage{
create_aneel_perdas_vulnerability(aneel_perdas_wider)
}
\arguments{
\item{aneel_perdas_wider}{Aneel - Relatório de Perdas de Energia compared
transformed in wider data version}
}
\value{
A dataframe
}
\description{
This function read the data from Aneel - Relatório de Perdas de Energia compared
wider data version and summarizes all vulnerabilities at UF, region, and country
view per stats.
}
\examples{
\dontrun{
  aneel_perdas_summarized_by_vulnerability <- create_aneel_perdas_vulnerability(
    aneel_perdas_wider
  )
}
}
