% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_anp_precoglp_vulnerability.R
\name{create_anp_precoglp_vulnerability}
\alias{create_anp_precoglp_vulnerability}
\title{Creates ANP - PRECOGLP Vulnerability}
\usage{
create_anp_precoglp_vulnerability(anp_precoglp_wider)
}
\arguments{
\item{anp_precoglp_wider}{ANP - Preço do GLP dataset transformed in wider data version}
}
\value{
A dataframe
}
\description{
This function read the data from ANP - Preço do GLP
wider data version and summarizes all vulnerabilities
at UF, region, and country view per stats.
}
\examples{
\dontrun{
  anp_preco_glp_wider <- anp_precoglp_transform_wider()
  anp_precoglp_summarized_by_vulnerability <- create_anp_precoglp_vulnerability(anp_precoglp_wider)
}
}
