% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mds_auxgas_vulnerability.R
\name{create_mds_auxgas_vulnerability}
\alias{create_mds_auxgas_vulnerability}
\title{Creates MDS - Auxílio Gás Vulnerability}
\usage{
create_mds_auxgas_vulnerability(mds_auxgas_wider)
}
\arguments{
\item{mds_auxgas_wider}{MDS - Auxílio Gás transformed in wider data version.}
}
\value{
A dataframe
}
\description{
This function read the data from MDS - Auxílio Gás and summarizes all vulnerabilities
at UF, region, and country view per stats.
}
\examples{
\dontrun{
  mds_auxgas_summarized_by_vulnerability <- create_mds_auxgas_vulnerability(
    dir = "data_mds_auxgas"
  )
}
}
