% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creates_longer_dataset.R
\name{creates_longer_dataset}
\alias{creates_longer_dataset}
\title{Create Longer Dataset}
\usage{
creates_longer_dataset(
  indicadores = list.files("ETL_pipeline/data/data-output/microdados-wider-rds",
    full.names = T, pattern = ".rds"),
  export_path = "ETL_pipeline/data/data-output/",
  lista_indicadores = list(pnad2016 = bidexpansaoenergetica::pnad2016, pnad2017 =
    bidexpansaoenergetica::pnad2017, pnad2018 = bidexpansaoenergetica::pnad2018, pnad2019
    = bidexpansaoenergetica::pnad2019, pnad2022 = bidexpansaoenergetica::pnad2022,
    pof2009 = bidexpansaoenergetica::pof2009, pof2018 = bidexpansaoenergetica::pof2018)
)
}
\arguments{
\item{indicadores}{A character vector of file paths pointing to `.rds` files
containing wide-format indicator datasets. Default is set to list all `.rds` files
in the `ETL/output/microdados_wider_rds` directory.}

\item{export_path}{A character string specifying the directory where the
reshaped datasets should be saved. Default is `"ETL/output/"`.}

\item{lista_indicadores}{a named list with indicators defined in bidexpansaoenergetica data export package, the usual is datasource+year}
}
\value{
None. The function saves the reshaped datasets as `.csv` files
in a `longer_csv` subdirectory within the specified `export_path`.
}
\description{
Processes and reshapes multiple indicator datasets into a standardized,
long-format dataset, saving the results in the specified export directory.
}
\details{
This function loads each specified `.rds` file containing indicator data,
extracts metadata (database name and year), and reshapes the data into a
long format. The function then saves each reshaped dataset in `.csv` format
in the `longer_csv` subdirectory within `export_path`.
}
\examples{
\dontrun{
creates_longer_dataset(
  indicadores = list.files('ETL/output/microdados_wider_rds',
                           full.names = TRUE, pattern = ".rds"),
  export_path = "ETL/output/"
)
}

}
